+++ CHANGELOG.md
@@ -6,9 +6,16 @@ This project adheres to [Semantic Versioning][semver].
 The format of this changelog is [a variant][lib9-versionning] of [Keep a Changelog][keep-changelog].
 New entries must be placed in a section entitled `Unreleased`.
 
+## Unreleased
+
+- Improve the command line help message.
+
+  Hopefully, this will make the provided options more explanatory.
+  Also, we improved the [documentation about type correspondence between BARE and TypeScript](./docs/types.md).
+
 ## 0.17.1 (2025-10-12)
 
-- Fix CLI argument parsing when using `compile` command to properly identify schema files
+- Fix CLI argument parsing when using `compile` command to properly identify schema files.
 
 ## 0.17.0 (2025-05-03)